requirements.txt
langchain 
markdown2 
langchain_openai 
langchain_experimental 
hnswlib 
accelerate 
optimum 
auto-gptq 
python-magic 
sentence_transformers